+++ packages/angular-radix/jest.config.js
@@ -22,6 +22,7 @@ module.exports = {
         baseUrl: '.',
         paths: {
           '@alaarab/ogrid-core': ['../core/src/index.ts'],
+          '@alaarab/ogrid-core/formula': ['../core/src/formula/index.ts'],
           '@alaarab/ogrid-core/testing': ['../core/src/testing/index.ts'],
           '@alaarab/ogrid-angular': ['../angular/src/index.ts'],
           '@alaarab/ogrid-angular/testing': ['../angular/src/testing/index.ts'],
@@ -31,6 +32,7 @@ module.exports = {
   },
   moduleNameMapper: {
     '^@alaarab/ogrid-core$': '<rootDir>/../core/src/index.ts',
+    '^@alaarab/ogrid-core/formula$': '<rootDir>/../core/src/formula/index.ts',
     '^@alaarab/ogrid-core/testing$': '<rootDir>/../core/src/testing/index.ts',
     '^@alaarab/ogrid-angular$': '<rootDir>/../angular/src/index.ts',
     '^@alaarab/ogrid-angular/testing$': '<rootDir>/../angular/src/testing/index.ts',
