+++ packages/angular-radix/tsup.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
   dts: false,
   target: 'es2020',
   minify: true,
+  external: ['@alaarab/ogrid-angular', '@alaarab/ogrid-core', '@alaarab/ogrid-core/formula', '@angular/core', '@angular/common', '@angular/cdk'],
   esbuildPlugins: [sassPlugin({ type: 'local-css' })],
   outExtension: () => ({ js: '.js' }),
 });
