+++ packages/angular/src/index.ts
@@ -131,13 +131,8 @@ export {
   validateRowIds,
   validateVirtualScrollConfig,
   indexToColumnLetter,
+  columnLetterToIndex,
   formatCellReference,
-  extractFormulaReferences,
-  processFormulaBarCommit,
-  deriveFormulaBarText,
-  handleFormulaBarKeyDown,
-  canInsertReference,
-  insertReferenceAtCursor,
   getResponsiveHiddenColumns,
   RESPONSIVE_BREAKPOINTS,
   resolveResponsiveConfig,
@@ -157,12 +152,9 @@ export {
   PEOPLE_SEARCH_DEBOUNCE_MS,
   SIDEBAR_TRANSITION_MS,
   Z_INDEX,
-  FORMULA_REF_COLORS,
-  FORMULA_BAR_CSS,
-  FORMULA_BAR_STYLES,
 } from '@alaarab/ogrid-core';
 
-// Core formula system
+// Formula system (from @alaarab/ogrid-core/formula subpath)
 export {
   FormulaError,
   FormulaEngine,
@@ -171,7 +163,6 @@ export {
   tokenize,
   parse,
   createBuiltInFunctions,
-  columnLetterToIndex,
   parseCellRef,
   parseRange,
   formatAddress,
@@ -190,7 +181,16 @@ export {
   CIRC_ERROR,
   GENERAL_ERROR,
   NA_ERROR,
-} from '@alaarab/ogrid-core';
+  extractFormulaReferences,
+  processFormulaBarCommit,
+  deriveFormulaBarText,
+  handleFormulaBarKeyDown,
+  canInsertReference,
+  insertReferenceAtCursor,
+  FORMULA_REF_COLORS,
+  FORMULA_BAR_CSS,
+  FORMULA_BAR_STYLES,
+} from '@alaarab/ogrid-core/formula';
 
 // Angular-specific types
 export type {
