+++ packages/angular/tsup.config.ts
@@ -10,5 +10,6 @@ export default defineConfig({
   dts: false,
   target: 'es2020',
   minify: true,
+  external: ['@alaarab/ogrid-core', '@alaarab/ogrid-core/formula', '@angular/core', '@angular/common'],
   outExtension: () => ({ js: '.js' }),
 });
