+++ packages/core/src/utils/cellReference.ts
@@ -23,6 +23,22 @@ export function indexToColumnLetter(index: number): string {
 }
 
 /**
+ * Parse a column letter string to a 0-based index.
+ * @example columnLetterToIndex('A')  // 0
+ * @example columnLetterToIndex('Z')  // 25
+ * @example columnLetterToIndex('AA') // 26
+ * @example columnLetterToIndex('AZ') // 51
+ */
+export function columnLetterToIndex(letters: string): number {
+  let result = 0;
+  const upper = letters.toUpperCase();
+  for (let i = 0; i < upper.length; i++) {
+    result = result * 26 + (upper.charCodeAt(i) - 64);
+  }
+  return result - 1;
+}
+
+/**
  * Format a cell reference string from a 0-based column index and a 1-based row number.
  * @example formatCellReference(0, 1)    // 'A1'
  * @example formatCellReference(2, 15)   // 'C15'
