+++ packages/core/src/utils/index.ts
@@ -97,7 +97,7 @@ export { applyFillValues } from './fillHelpers';
 export type { IFillFormulaOptions } from './fillHelpers';
 export { UndoRedoStack } from './undoRedoStack';
 export { validateColumns, validateRowIds, validateVirtualScrollConfig } from './validation';
-export { indexToColumnLetter, formatCellReference } from './cellReference';
+export { indexToColumnLetter, columnLetterToIndex, formatCellReference } from './cellReference';
 export { extractFormulaReferences, processFormulaBarCommit, deriveFormulaBarText, handleFormulaBarKeyDown, canInsertReference, insertReferenceAtCursor } from './formulaBarHelpers';
 export type { FormulaReference } from './formulaBarHelpers';
 export { getResponsiveHiddenColumns, RESPONSIVE_BREAKPOINTS, resolveResponsiveConfig, applyResponsiveHiding } from './responsiveColumns';
