+++ packages/core/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/formula/index.ts'],
   format: ['esm'],
   outDir: 'dist/esm',
   splitting: false,
