+++ packages/js/src/index.ts
@@ -174,13 +174,8 @@ export {
   validateRowIds,
   validateVirtualScrollConfig,
   indexToColumnLetter,
+  columnLetterToIndex,
   formatCellReference,
-  extractFormulaReferences,
-  processFormulaBarCommit,
-  deriveFormulaBarText,
-  handleFormulaBarKeyDown,
-  canInsertReference,
-  insertReferenceAtCursor,
   getResponsiveHiddenColumns,
   RESPONSIVE_BREAKPOINTS,
   resolveResponsiveConfig,
@@ -200,12 +195,9 @@ export {
   PEOPLE_SEARCH_DEBOUNCE_MS,
   SIDEBAR_TRANSITION_MS,
   Z_INDEX,
-  FORMULA_REF_COLORS,
-  FORMULA_BAR_CSS,
-  FORMULA_BAR_STYLES,
 } from '@alaarab/ogrid-core';
 
-// Core formula system
+// Formula system (from @alaarab/ogrid-core/formula subpath)
 export {
   FormulaError,
   FormulaEngine,
@@ -214,7 +206,6 @@ export {
   tokenize,
   parse,
   createBuiltInFunctions,
-  columnLetterToIndex,
   parseCellRef,
   parseRange,
   formatAddress,
@@ -233,7 +224,16 @@ export {
   CIRC_ERROR,
   GENERAL_ERROR,
   NA_ERROR,
-} from '@alaarab/ogrid-core';
+  extractFormulaReferences,
+  processFormulaBarCommit,
+  deriveFormulaBarText,
+  handleFormulaBarKeyDown,
+  canInsertReference,
+  insertReferenceAtCursor,
+  FORMULA_REF_COLORS,
+  FORMULA_BAR_CSS,
+  FORMULA_BAR_STYLES,
+} from '@alaarab/ogrid-core/formula';
 
 // Shadow core column types with vanilla JS extensions
 export type { IColumnDef, IColumnGroupDef, ICellEditorContext } from './types/columnTypes';
