+++ packages/react-fluent/tsup.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
   dts: false,
   target: 'es2020',
   minify: true,
+  external: ['@alaarab/ogrid-react', '@alaarab/ogrid-core', '@alaarab/ogrid-core/formula', '@tanstack/react-virtual', '@fluentui/react-components', '@fluentui/react-icons', 'react', 'react-dom'],
   esbuildOptions(options) {
     options.jsx = 'automatic';
     options.banner = { js: "import './index.css';" };
