+++ packages/react-radix/tsup.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
   dts: false,
   target: 'es2020',
   minify: true,
+  external: ['@alaarab/ogrid-react', '@alaarab/ogrid-core', '@alaarab/ogrid-core/formula', '@tanstack/react-virtual', '@radix-ui/react-checkbox', '@radix-ui/react-dropdown-menu', '@radix-ui/react-popover', 'react', 'react-dom'],
   esbuildOptions(options) {
     options.jsx = 'automatic';
     options.banner = { js: "import './index.css';" };
