+++ packages/vue-primevue/tsup.config.ts
@@ -10,5 +10,6 @@ export default defineConfig({
   dts: false,
   target: 'es2020',
   minify: true,
+  external: ['@alaarab/ogrid-vue', '@alaarab/ogrid-core', '@alaarab/ogrid-core/formula', 'vue', 'primevue'],
   outExtension: () => ({ js: '.js' }),
 });
