+++ packages/vue/src/components/FormulaRefOverlay.ts
@@ -6,7 +6,7 @@
  */
 
 import { defineComponent, h, ref, watch, type PropType } from 'vue';
-import { FORMULA_REF_COLORS, type FormulaReference } from '@alaarab/ogrid-core';
+import { FORMULA_REF_COLORS, type FormulaReference } from '@alaarab/ogrid-core/formula';
 
 interface RefRect {
   top: number;
